+++ vesting-solution/hie.yaml
@@ -2,5 +2,3 @@ cradle:
   cabal:
     - path: "./vesting-contract/src"
       component: "vesting-contract:lib:vesting-contract"
-    - path: "./vesting-contract/pab"
-      component: "vesting-contract:exe:plutus-starter-pab"
