+++ vesting-solution/v1-vesting-contract/build-project.sh
@@ -1,0 +1,3 @@
+cabal build -w ghc-8.10.4
+cabal run vesting-contract
+echo "DONE"
