+++ vesting-solution/v1-vesting-contract/full-build-project.sh
@@ -1,0 +1,5 @@
+cabal clean
+cabal update
+cabal build -w ghc-8.10.4 -O2
+cabal run vesting-contract
+echo "DONE"
