+++ vesting-solution/v1-vesting-contract/quick-build.sh
@@ -1,0 +1,2 @@
+cabal build -w ghc-8.10.4
+echo "DONE!"
