+++ vesting-solution/v1-vesting-contract/src/CheckFuncs.hs
