+++ vesting-solution/v1-vesting-contract/src/DataTypes.hs
