+++ vesting-solution/vesting-contract/quick-build.sh
@@ -1,2 +1,1 @@
-cabal build -w ghc-8.10.4
-echo "DONE!"
+cabal build -w ghc-8.10.7
