+++ vesting-solution/vesting-contract/validator.addr
@@ -1,0 +1,1 @@
+addr_test1wqvfxsx9cnudnw4p2zds7dd35yfncy8avah68jejk2jlmgsc4knz9
