+++ vesting-solution/vesting-contract/validator.bytes
@@ -1,0 +1,1 @@
+[24, 147, 64, 197, 196, 248, 217, 186, 161, 80, 155, 15, 53, 177, 161, 19, 60, 16, 253, 103, 111, 163, 203, 50, 178, 165, 253, 162]
