+++ vesting-solution/vesting-contract/validator.hash
@@ -1,0 +1,1 @@
+189340c5c4f8d9baa1509b0f35b1a1133c10fd676fa3cb32b2a5fda2
